requirements.txt
Django==4.0.2
django-extensions==3.1.5
django-debug-toolbar==3.2.4
pillow==9.0.1
beautifulsoup4==4.10.0
django-environ==0.8.1
django-allauth==0.48.0
django-mptt == 0.13.4
ipython==8.1.0
